Promoted functionality from rundemos.clj into demos/ so that it can be re-compiled when connected to swank.
rundemos.clj is a bootstrapper and not part of the swank classpath. It should
not contain much logic. Any logic it does contain cannot be re-compiled when
connecting to swank from an editor so it makes sense to move the logic to a source area that is in the classpath.