Got the basic-auth demo working.
This required quite a few changes to the architecture. Firstly, a plugin init
function was required to save the web-namespaces into the state such that they
could be queried later on. Instead of querying the web functions during the
resource-exists? step, this is now done at the end of the status calculation
phase. This means that the state can also be used in the qualification of web
functions.

The additional feature that this change made possible is shown with the
status-views demo. A page can declare a status and then it is used as a default
page for that status.